+++ drift/lib/src/runtime/api/connection.dart
@@ -20,11 +20,22 @@ class DatabaseConnection implements QueryExecutor {
 
   /// Constructs a raw database connection from the [executor] and optionally a
   /// specified [streamQueries] implementation to use.
+  ///
+  /// When creating a default stream query implementation,
+  /// [closeStreamsSynchronously] determines whether drift stops query streams
+  /// immediately after the last listener detaches. By default, drift will wait
+  /// for one event loop iteration to avoid duplicate work for e.g.
+  /// `StreamBuilder` setups that reconnect on rebuilds. Enabling that option
+  /// may be useful in test setups that throw exceptions for timers persisting
+  /// after tests.
   DatabaseConnection(
     this.executor, {
     StreamQueryStore? streamQueries,
     this.connectionData,
-  }) : streamQueries = streamQueries ?? StreamQueryStore();
+    bool closeStreamsSynchronously = false,
+  }) : streamQueries = streamQueries ??
+            StreamQueryStore(
+                closeStreamsSynchronously: closeStreamsSynchronously);
 
   /// Constructs a [DatabaseConnection] from the [QueryExecutor] by using the
   /// default type system and a new [StreamQueryStore].