+++ drift/test/integration_tests/cancellation_test.dart
@@ -2,10 +2,14 @@
 @TestOn('vm')
 library;
 
+import 'dart:async';
+
 import 'package:drift/drift.dart';
 import 'package:drift/isolate.dart';
+import 'package:drift/native.dart';
 import 'package:test/test.dart';
 
+import '../generated/todos.dart';
 import '../test_utils/database_vm.dart';
 import 'cancellation_test_support.dart';
 
@@ -48,4 +52,33 @@ void main() {
       await runTest(db);
     });
   }, skip: 'todo: Cancellations are currently broken on Dart 2.15');
+
+  test('can cancel streams synchronously', () async {
+    final createdTimers = <Timer>[];
+
+    await runZoned(
+      () async {
+        final database = TodoDb(DatabaseConnection(
+          NativeDatabase.memory(),
+          closeStreamsSynchronously: true,
+        ));
+
+        await database.todosTable.all().watch().first;
+        // This cancels a stream subscription - drift would usually set up a
+        // timer to wait for that.
+      },
+      zoneSpecification: ZoneSpecification(
+        createTimer: (self, parent, zone, duration, f) {
+          final timer = parent.createTimer(zone, duration, f);
+          createdTimers.add(timer);
+          return timer;
+        },
+      ),
+    );
+
+    expect(
+        createdTimers,
+        everyElement(
+            isA<Timer>().having((e) => e.isActive, 'isActive', isFalse)));
+  });
 }